Added new platform type for .NET Core 3.1, selected by the JIT GUID
